functions/requirements.txt
firebase-functions
firebase-admin
google-cloud-vision
google-api-python-client